Deal with roadblock avoid. + stub-only at run time
And make the single usage function validate_extension static
